crypto/requirements.txt
pandas==1.2.4
python-binance==1.0.14
cryptography==3.4.7
cx-Oracle==8.3.0